package/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "tue-api-wrapper"
version = "0.2.0"
description = "Unified Alma and ILIAS client plus API for the University of Tuebingen."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12,<5",
  "fastapi>=0.115,<1",
  "python-dateutil>=2.9,<3",
  "requests>=2.32,<3",
  "uvicorn>=0.34,<1",
]

[project.scripts]
alma-academics = "tue_api_wrapper.alma_academics_cli:main"
alma-timetable = "tue_api_wrapper.cli:main"
alma-document = "tue_api_wrapper.alma_documents_cli:main"
ilias-content = "tue_api_wrapper.ilias_content_cli:main"
ilias-learning = "tue_api_wrapper.ilias_learning_cli:main"
ilias-root = "tue_api_wrapper.ilias_cli:main"
tue-api-server = "tue_api_wrapper.api_server:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
